+++ Dockerfile
@@ -4,7 +4,7 @@ FROM golang:1.19.1 as builder
 WORKDIR /workspace
 
 # Copy the Go Modules manifests
-COPY go.mod go.mod go.sum go.sum ./
+COPY go.mod go.sum ./
 
 # cache deps before building and copying source so that we don't need to re-download as much
 # and so that source changes don't invalidate our downloaded layer
