+++ pkg/keda/keda_test.go
@@ -1,4 +1,4 @@
-package controllers
+package keda
 
 import (
 	"testing"
@@ -66,7 +66,7 @@ func TestHasExistingKedaInstallation(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := HasExistingKedaInstallation(tt.c, logger)
+			got, err := IsInstalled(tt.c, logger)
 			if (err != nil) != tt.wantErr {
 				t.Errorf("HasExistingKedaInstallation() error = %v, wantErr %v", err, tt.wantErr)
 				return
